requirements.txt
djangorestframework
django-cors-headers
pillow
djoser
pytils
django-cleanup
django-ckeditor


